pyproject.toml
[tool.poetry]
name = "osu-dreamer"
version = "5.0"
description = "a diffusion-based ML model for generating osu! maps from raw audio"
authors = ["jason <jason@jwon.xyz>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
beartype = "^0.18.3"
jaxtyping = "^0.2.28"
pytorch-lightning = "^2.2.2"
jsonargparse = {extras = ["signatures"], version = "^4.28.0"}
librosa = "^0.10.1"
bezier = "^2023.7.28"
einops = "^0.7.0"
tensorboard = "^2.16.2"
matplotlib = "^3.8.4"
ipykernel = "^6.29.4"
mutagen = "^1.47.0"
click = "^8.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
